move vide swin layer to model specific directory
